src/jetson/requirements.txt
absl-py==0.8.0
apt-clone==0.2.1
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.0
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.6.0
bleach==3.1.0
blinker==1.4
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
decorator==4.4.0
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
entrypoints==0.3
feedparser==5.2.1
gast==0.2.2
graphsurgeon==0.4.1
grpcio==1.23.0
h5py==2.9.0
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.1
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mock==3.0.5
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.1
oauth==1.0.1
oauthlib==2.0.6
PAM==0.4.2
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.1
ptyprocess==0.6.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.4.2
pygobject==3.26.1
PyICU==1.9.8
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.2
pyRFC3339==1.0
pyrsistent==0.15.4
python-apt==1.6.4
python-dateutil==2.8.0
python-debian==0.1.32
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
pyzmq==18.1.0
qtconsole==4.5.4
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.13.2
six==1.12.0
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1+nv19.4
tensorrt==5.1.6.1
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
torch==1.1.0a0+b457266
torchvision==0.3.0
tornado==6.0.3
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
uff==0.6.3
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.22
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.5.1
xkit==0.0.0
zope.interface==4.3.2
